Handle case of subsequent dummy characters
Previously this resulted in spaces getting output because we didn't
remove the old dummy character. Note that we still need this dummy
character to read the current styles, so we first remove it *after*
reading its attributes.

This also changes the dummy character from a space to @ to make it
easier to debug.